requirements.txt
haystack-ai
python-dotenv
sentence-transformers
openai
qdrant-haystack
transformers
torch
PyPDF2 